Implemented deactivate and delete user Commented urls.py extensively Modified .gitignore, now ignoring .pyc and .sqlite3.

+++ lfc-api/LookingForConcerts/lfc_backend/views.py
@@ -6,7 +6,6 @@ from lfc_backend.models import RegisteredUser, Concert, Tag, Report, Location, R
 from lfc_backend.serializers import ConcertSerializer,LocationSerializer, RegisteredUserSerializer, CommentSerializer, RatingSerializer, ImageSerializer, ArtistSerializer
 
 from django.contrib.auth import authenticate, login # for user authentication and login
-from django.contrib.auth import logout # for user logout
 from django.contrib.auth.decorators import login_required, permission_required # permissions
 from django.shortcuts import render, redirect
 from django.core.exceptions import ObjectDoesNotExist
@@ -18,6 +17,14 @@ from spotipy.oauth2 import SpotifyClientCredentials
 import requests
 import json
 import re
+
+from rest_framework_simplejwt.tokens import RefreshToken # for blacklisting tokens upon user logout
+
+from rest_framework_simplejwt.views import (
+    TokenObtainPairView,
+    TokenRefreshView,
+    TokenVerifyView,
+)
 
 # The actual python functions that do the backend work.
 
@@ -35,7 +42,7 @@ def list_users(request):
     serializer = RegisteredUserSerializer(registered_users, many=True)
     return Response(serializer.data)
 
-@api_view(['GET'])
+@api_view(['POST'])
 def follow_user(request,pk):
     '''
     follows the registered user with primary key value pk.
@@ -50,7 +57,7 @@ def follow_user(request,pk):
     me.following.add(following_user)
     return Response(status = status.HTTP_200_OK)
 
-@api_view(['GET'])
+@api_view(['POST'])
 def unfollow_user(request,pk):
     '''
     unfollows the registered user with primary key value pk.
@@ -65,15 +72,6 @@ def unfollow_user(request,pk):
     me.following.remove(following_user)
     return Response(status = status.HTTP_200_OK)
 
-@api_view(['GET'])
-def user_detail(request,pk):
-    '''
-    returns a specific registered user
-    '''
-    registered_user = RegisteredUser.objects.get(pk=pk)
-    serializer = RegisteredUserSerializer(registered_user)
-    return Response(serializer.data)
-
 @api_view(['POST'])
 def signup(request):
     '''
@@ -81,25 +79,18 @@ def signup(request):
     '''
     serializer = RegisteredUserSerializer(data = request.data) # create the user
     if serializer.is_valid():
+        # MIGHT ADD SOME REQUIREMENTS FOR PASSWORD.
+        # E.G. SHOULD CONTAIN AT LEAST A NUMBER, A CAPITAL AND SMALL LETTER ETC.
         # SIGNUP
         registered_user = serializer.save() # save the user to the database
-        # LOGIN
-        username = registered_user.username # username
-        password = request.data['password'] # unhashed password that the user entered
-
-        # MIGHT ADD SOME REQUIREMENTS FOR PASSWORD. E.G. SHOULD CONTAIN AT LEAST A NUMBER, A CAPITAL AND SMALL LETTER ETC.
-
-        user = authenticate(username=username, password=password) # authenticate() hashes the given function inside before checking
-        login(request,user) # log in the user right after signup
         return Response(serializer.data, status = status.HTTP_201_CREATED) # success!
     else:
         return Response(serializer.errors, status = status.HTTP_400_BAD_REQUEST) # something went wrong!
 
 @api_view(['DELETE'])
-def delete_user(request):
+def deactivate_user(request):
     '''
-    deletes the user
-    actually, deactivates his account.
+    deactivates the account of this user
     '''
     if (request.user.is_authenticated):
         pk = request.user.id
@@ -112,14 +103,27 @@ def delete_user(request):
     except:
         return Response(status = status.HTTP_404_NOT_FOUND)
 
-    print(user.email)
-
     user.is_active=False
     user.save(update_fields=['is_active'])
+    return Response({"message": "Deactivated user " + str(username)},status = status.HTTP_204_NO_CONTENT)
 
-    #user.delete()
+@api_view(['DELETE'])
+def delete_user(request):
+    '''
+    deletes the user with the given username.
+    only admins are authorized.
+    '''
+    if (request.user.is_staff == False):
+        return Response(status=status.HTTP_401_UNAUTHORIZED)
+    else:
+        try:
+            username = request.data['username']
+            user = RegisteredUser.objects.get(username=username)
+        except:
+            return Response(status = status.HTTP_404_NOT_FOUND)
 
-    return Response({"message": "Deleted User" + str(username)},status = status.HTTP_204_NO_CONTENT)
+        user.delete()
+        return Response({"message": "Deleted user " + str(username)},status = status.HTTP_204_NO_CONTENT)
 
 @api_view(['DELETE'])
 @permission_required
@@ -127,13 +131,16 @@ def delete_all_users(request):
     '''
     deletes all users
     '''
-    print("Deleting all users...")
-    try:
-        users = RegisteredUser.objects.all()
-    except ObjectDoesNotExist:
-        return Response(status = status.HTTP_404_NOT_FOUND)
-    users.delete()
-    return Response(status = status.HTTP_204_NO_CONTENT)
+    if (request.user.is_staff == False):
+        return Response(status=status.HTTP_401_UNAUTHORIZED)
+    else:
+        print("Deleting all users...")
+        try:
+            users = RegisteredUser.objects.all()
+        except ObjectDoesNotExist:
+            return Response(status = status.HTTP_404_NOT_FOUND)
+        users.delete()
+        return Response(status = status.HTTP_204_NO_CONTENT)
 
 @api_view(['GET'])
 def get_user_info(request):
@@ -143,24 +150,26 @@ def get_user_info(request):
         serializer = RegisteredUserSerializer(request.user)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
-@api_view(['POST'])
-def registered_user_logout(request):
+@api_view(['GET'])
+def get_user_concerts(request):
     '''
-    logs out the user
+    returns all the concerts of a user
     '''
-    logout(request)
-    print("Logged out.")
-    return Response(status=status.HTTP_200_OK)
-    # Redirect to a success page.
+    user = request.user
+    if user.is_authenticated:
+        concerts=user.concerts.all()
+        serializer = ConcertSerializer(concerts)
+        return Response(serializer.data,status=status.HTTP_200_OK)
+    else:
+        return Response(status=status.HTTP_401_UNAUTHORIZED)
 
 '''
 CONCERT FUNCTIONS
 '''
-
 @api_view(['GET'])
 def list_concerts(request):
     '''
-    returns all concerts
+    returns all concerts in the system
     @params: None
     '''
     if request.method =='GET':
@@ -171,8 +180,6 @@ def list_concerts(request):
         return status.HTTP_400_BAD_REQUEST
 
 @api_view(['POST'])
-#@login_required()
-#@permission_required('IsAuthenticated')
 def create_concert(request):
     '''
     inserts a concert into the database
@@ -217,7 +224,7 @@ def create_concert(request):
         return Response(serializer.data, status = status.HTTP_201_CREATED)
     return Response(serializer.errors, status = status.HTTP_400_BAD_REQUEST)
 
-@api_view(['POST'])
+@api_view(['GET'])
 def search_concerts(request):
     '''
     searches the post data with concerts name, location, artist and tags
@@ -265,7 +272,7 @@ def concert_detail(request, pk):
         concert.delete()
         return Response(status = status.HTTP_204_NO_CONTENT)
 
-@api_view(['GET'])
+@api_view(['POST'])
 def subscribe_concert(request, pk):
     '''
     adds user to concerts user list
@@ -280,7 +287,7 @@ def subscribe_concert(request, pk):
     concert.attendees.add(request.user)
     return Response(status  = status.HTTP_200_OK)
 
-@api_view(['GET'])
+@api_view(['POST'])
 def unsubscribe_concert(request, pk):
     '''
     removes user from concerts user list
@@ -299,7 +306,7 @@ def unsubscribe_concert(request, pk):
 '''
 ARTIST FUNCTIONS
 '''
-@api_view(['POST'])
+@api_view(['GET'])
 def search_artists(request):
     data = request.data
     client_credentials_manager = SpotifyClientCredentials(client_id='60ab66df7413492bbc86150d7a3617d7', client_secret='007ccb30ee7e4eb98478b7a34fc869e4')
@@ -334,7 +341,6 @@ def location_detail(request, pk):
         location = Location.objects.get(pk=pk) # <-- this pk might be location_id, I'm not sure. <--- It is
     except:
         return Response(status = status.HTTP_404_NOT_FOUND)
-    # returns the concert with the given primary key
     serializer = LocationSerializer(location)
     return Response(serializer.data)
 
@@ -344,6 +350,9 @@ RATING FUNCTIONS
 
 @api_view(['POST'])
 def rate_concert(request,pk):
+    '''
+    Adds ratings by the logged in user to the concert with given pk
+    '''
     if (not request.user.is_authenticated):
         return Response(status=status.HTTP_401_UNAUTHORIZED)
 
@@ -382,7 +391,7 @@ TAG FUNCTIONS
 def get_tags(request, search_str):
     if (not request.user.is_authenticated):
         return Response({'Error':'User is not authenticated'},status=status.HTTP_401_UNAUTHORIZED)
-    
+
     API_ENDPOINT = "https://www.wikidata.org/w/api.php"
     query = search_str
     params = {
@@ -392,7 +401,7 @@ def get_tags(request, search_str):
         'search' : query,
         'type':'item'
     }
-    
+
     r = requests.get(API_ENDPOINT, params = params)
     json_response = r.json()['search']
     lenght =  len(json_response)
@@ -404,7 +413,7 @@ def get_tags(request, search_str):
                 value = json_response[i]['label']
                 context = json_response[i]['description']
                 t = '{"value":"'+value.replace('"','')+ '","context":"'+context.replace('"','')+'"}'
-                print(t)    
+                print(t)
                 tags.append(json.loads(t))
 
     return Response(tags, status.HTTP_200_OK)
@@ -415,6 +424,9 @@ COMMENT FUNCTIONS
 
 @api_view(['POST'])
 def create_comment(request,pk):
+    '''
+    Adds a comment by the logged in user to the concert with given pk
+    '''
     if (not request.user.is_authenticated):
         return Response(status=status.HTTP_401_UNAUTHORIZED)
 
@@ -425,7 +437,6 @@ def create_comment(request,pk):
 
     serializer = CommentSerializer(data = request.data)
     serializer.is_valid()
-    #comment = serializer.save(owner = request.user)
     comment = serializer.save()
     request.user.comments.add(comment)
     concert.comments.add(comment)
